+++ changelog.md
@@ -5,6 +5,10 @@
 - Run Composer Install
 - Run Migrate
 
+1.13.3 **Calendar Dark & Date Picker**
+- Calendar dark mode
+- Change date input with datepicker
+
 1.13.2 **Passing Script, Send WA, Search after CU**
 - Add base url passing to script tag
 - Add send to WhatsApp on customer and job table